Got match data storage working
Backend now stores data from the website upon submission. The system uses fake event and team ids for testing purposes. There is a specific line of code in scout.go that creates a match upon submission. This should be uncommented once to create the match for the rest of the system to work with. Failing to do so will result in errors saying the system cannot find the specified match.